replace a racy usage of WaitGroups with channels
According to the [documentation](https://golang.org/pkg/sync/#WaitGroup), wait
groups' `Add` and `Done` have to be synchronized. Currently the http server
will call `Add` in a the http server `ConnState` callback and `Done` in the
runner's `Run` method. This will cause a race condition, for example in [this
inigo
build](https://diego.ci.cf-app.com/teams/main/pipelines/main/jobs/inigo-mysql/builds/412).

The new implementation uses a channel and a counter which is shared between the
`ConnState` callback and the runner's `Run` method.

[#147374047]

Signed-off-by: John Shahid <jvshahid@gmail.com>